# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MovieFlix API",
    "version": "1.0.0",
    "description": "API responsável por gerenciar os filmes do MovieFlix."
  },
  "paths": {
    "/movies": {
      "get": {
        "tags": ["Movie"],
        "summary": "Retorna uma lista de filmes",
        "responses": {
          "200": {
            "description": "Lista de filmes retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}": {
      "put": {
        "tags": ["Movie"],
        "summary": "Atualizar um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do filme a ser atualizado",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Filme atualizado com sucesso"
          },
          "404": {
            "description": "Filme não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "tags": ["Movie"],
        "summary": "Remover um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do filme a ser removido",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filme removido com sucesso"
          },
          "404": {
            "description": "Filme não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/movies/{genreName}": {
      "get": {
        "tags": ["Movie"],
        "summary": "Filtrar filmes por gênero",
        "parameters": [
          {
            "name": "genreName",
            "in": "path",
            "required": true,
            "description": "Nome do gênero para filtrar os filmes",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filmes filtrados por gênero retornados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/genres/{id}": {
      "put": {
        "tags": ["Genre"],
        "summary": "Atualizar um gênero",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do gênero a ser atualizado",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenreInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gênero atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genre"
                }
              }
            }
          },
          "400": {
            "description": "O nome do gênero é obrigatório"
          },
          "404": {
            "description": "Gênero não encontrado"
          },
          "409": {
            "description": "Este nome de gênero já existe"
          },
          "500": {
            "description": "Falha ao atualizar o gênero. Erro interno do servidor"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do filme"
          },
          "title": {
            "type": "string",
            "description": "Título do filme"
          },
          "genre_id": {
            "type": "integer",
            "description": "Gênero do filme"
          },
          "genres": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "ID do gênero"
              },
              "name": {
                "type": "string",
                "description": "Nome do gênero"
              }
            }
          },
          "language_id": {
            "type": "integer",
            "description": "Idioma do filme"
          },
          "languages": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "ID do idioma"
              },
              "name": {
                "type": "string",
                "description": "Nome do idioma"
              }
            }
          },
          "oscar_count": {
            "type": "integer",
            "description": "Número de Oscars recebidos pelo filme"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "description": "Data de lançamento do filme"
          }
        }
      },
      "MovieInput": {
        "type": "object",
        "required": [
          "title",
          "genre_id",
          "language_id",
          "oscar_count",
          "release_date"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Título do filme"
          },
          "genre_id": {
            "type": "integer",
            "description": "ID do gênero do filme"
          },
          "language_id": {
            "type": "integer",
            "description": "ID do idioma do filme"
          },
          "oscar_count": {
            "type": "integer",
            "description": "Número de Oscars recebidos pelo filme"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "description": "Data de lançamento do filme"
          }
        }
      },
      "Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do gênero"
          },
          "name": {
            "type": "string",
            "description": "Nome do gênero"
          }
        }
      },
      "GenreInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do gênero"
          }
        }
      }
    }
  }
}